+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+### Changed
+- Tests helpers are no longer in the Browser Kit testing style, and instead use a `TestResponse` class that extends
+the standard Laravel test response.
+- The `InteractsWithResources` test helper trait has been merged into `MakesJsonApiRequests`.
+
 ### Fixed
 - [#66] Content negotiation no longer sends a `415` response if a request does not have body content.